logic: move to analyzer, rename to abiparse
