+++ analyzer/runtime/abiparse/abiparse_test.go
@@ -1,4 +1,4 @@
-package logic
+package abiparse
 
 import (
 	_ "embed"
@@ -29,10 +29,10 @@ func init() {
 	Varied = artifactVaried.ABI
 }
 
-func TestEVMParseTypes(t *testing.T) {
+func TestParseTypes(t *testing.T) {
 	data, err := hex.DecodeString("b23a194fffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffff0000000000000000000000000000000000000000000000000000000000000001ffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffff0000000000000000000000000000000000000000000000000000000000000001000000000000000000000000000000000000000000000000000000000000000101010101010101010101010101010101010101010101010101010101010101010000000000000000000000000101010101010101010101010101010101010101010101010101010101010101010101010101010102020202000000000000000000000000000000000000000000000000000000000000000000000000000000010000000000000000000000000000000000000000000000000000000000000001000000000000000000000000000000000000000000000000000000000000000100000000000000000000000000000000000000000000000000000000000000010000000000000000000000000000000000000000000000000000000000000220000000000000000000000000000000000000000000000000000000000000026000000000000000000000000000000000000000000000000000000000000002a000000000000000000000000000000000000000000000000000000000000002e00000000000000000000000000000000000000000000000000000000000000320000000000000000000000000000000000000000000000000000000000000000101000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000001610000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000010000000000000000000000000000000000000000000000000000000000000001000000000000000000000000000000000000000000000000000000000000000100000000000000000000000000000000000000000000000000000000000000010000000000000000000000000000000000000000000000000000000000000001000000000000000000000000000000000000000000000000000000000000004000000000000000000000000000000000000000000000000000000000000000016100000000000000000000000000000000000000000000000000000000000000")
 	require.NoError(t, err)
-	method, args, err := EVMParseData(data, Varied)
+	method, args, err := ParseData(data, Varied)
 	require.NoError(t, err)
 	require.Equal(t, Varied.Methods["test"], *method)
 	jsonExpected := []string{
@@ -60,11 +60,11 @@ func TestEVMParseTypes(t *testing.T) {
 	}
 }
 
-func TestEVMParseData(t *testing.T) {
+func TestParseData(t *testing.T) {
 	// https://explorer.emerald.oasis.dev/tx/0x1ac7521df4cda38c87cff56b1311ee9362168bd794230415a37f2aff3a554a5f/internal-transactions
 	data, err := hex.DecodeString("095ea7b3000000000000000000000000250d48c5e78f1e85f7ab07fec61e93ba703ae6680000000000000000000000000000000000000000000000003782dace9d900000")
 	require.NoError(t, err)
-	method, args, err := EVMParseData(data, evmabi.ERC20)
+	method, args, err := ParseData(data, evmabi.ERC20)
 	require.NoError(t, err)
 	require.Equal(t, evmabi.ERC20.Methods["approve"], *method)
 	require.Equal(t, []interface{}{
@@ -73,7 +73,7 @@ func TestEVMParseData(t *testing.T) {
 	}, args)
 }
 
-func TestEVMParseEvent(t *testing.T) {
+func TestParseEvent(t *testing.T) {
 	// https://explorer.emerald.oasis.dev/tx/0x1ac7521df4cda38c87cff56b1311ee9362168bd794230415a37f2aff3a554a5f/logs
 	topicsHex := []string{
 		"8c5be1e5ebec7d5bd14f71427d1e84f3dd0314c0f7b2291e5b200ac8c7c3b925",
@@ -88,7 +88,7 @@ func TestEVMParseEvent(t *testing.T) {
 	}
 	data, err := hex.DecodeString("0000000000000000000000000000000000000000000000003782dace9d900000")
 	require.NoError(t, err)
-	event, args, err := EVMParseEvent(topics, data, evmabi.ERC20)
+	event, args, err := ParseEvent(topics, data, evmabi.ERC20)
 	require.NoError(t, err)
 	require.Equal(t, evmabi.ERC20.Events["Approval"], *event)
 	require.Equal(t, []interface{}{
