+++ analyzer/runtime/abiparse/test_contracts/artifacts/Varied.json
